Rename scene_importer to importer
